==> code/00_libraries.R <==
# title: Packages to source into scripts
# author: "Sam Csik"
# date created: "2020-10-01"
# date edited: "2020-10-01"
# R version: 3.6.3
# input: "NA"
# output: "NA"

##############################
# Load packages
##############################

library(dataone)
library(tidyverse)
library(eatocsv) 
library(svMisc)
library(rvest)
library(xml2)
library(tidytext)
library(patchwork)
library(DT)
library(EML)
library(emld)
library(here)
library(lubridate)
library(janitor)
